% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EURdata2-data.R
\docType{data}
\name{EURdata2}
\alias{EURdata2}
\title{Arabidopsis QTL data on gravitropism}
\format{An object of class data.frame}
\usage{
data(EURdata2)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(EURdata2)
myVars <- EURdata2$snpID
}
\keyword{datasets}
