% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSNPs.R
\name{findSNPs}
\alias{findSNPs}
\title{Find the SNPs that match with the input SNP.}
\usage{
findSNPs(x, y)
}
\arguments{
\item{x}{Data frame with two columns (input_snp and matched_snps)}

\item{y}{Character with the input snp (chr:pos)}
}
\value{
Vector with all the matched SNPs ordered by chromosome and then by position.
}
\description{
Find the SNPs that match with the input SNP.
}
\examples{
SNP <- findSNPs(matched_snps, "1:1361641")

}
