% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSNPs2.R
\name{findSNPs2}
\alias{findSNPs2}
\title{Find the SNPs that match with the input SNP.}
\usage{
findSNPs2(x, y)
}
\arguments{
\item{x}{Data frame with two columns (input_snp and matched_snps)}

\item{y}{Character with the input snp (chr:pos)}
}
\value{
Vector with all the matched.
}
\description{
The same as "findSNPs" but without ordering the SNPs.
}
\examples{
SNP <- findSNPs2(matched_snps, "1:1361641")

}
