% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMatches.R
\name{listMatches}
\alias{listMatches}
\title{List the matches of each SNP.}
\usage{
listMatches(matched_snps)
}
\arguments{
\item{matched_snps}{Data frame with columns "input_snp" and "matched_snp".}
}
\value{
List with all the SNPs and a vector of its matches ordered by chromosome and position.
}
\description{
This function reads the output from "SNPmatch" and outputs the same
information as a list, where each object is a new input SNP which has
inside a vector with all its matches.
}
\examples{
SNP <- listMatches(matched_snps)

}
