% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderSNPs.R
\name{orderSNPs}
\alias{orderSNPs}
\title{Order a vector of SNPs}
\usage{
orderSNPs(SNPvector)
}
\arguments{
\item{SNPvector}{A character vector with the SNPs in format chr:pos.}
}
\value{
Vector with all SNPs ordered by chromosome and then by position.
}
\description{
Order a vector of SNPs
}
\examples{
myVector_ordered <- orderSNPs(myVector)

}
