% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permSNPs.R
\name{permSNPs}
\alias{permSNPs}
\title{Randomly choose SNPs with similar characteristics to your input SNPs.}
\usage{
permSNPs(SNPlist, numSNPs)
}
\arguments{
\item{SNPlist}{A list where each object is a input SNP (specified in the name of the object) with a vector of its matches.}

\item{numSNPs}{Numeric value with the information of the original number of SNPs.}
}
\value{
Vector with a randomly chosen number of variants with similar characteristics to the input SNPs.
}
\description{
Based on the information provided of the matches, the function chooses randomly
the matches with similar characteristics to the input SNPs. In case the number of original
SNPs (numSNPs) is not the same as the length of SNPlist, the difference will be filled with
random variants of the already sampled available SNPs.
}
\examples{
permVector <- permSNPs(SNP_list, 8000)

}
