% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPmatch_XList.R
\name{SNPmatch_XList}
\alias{SNPmatch_XList}
\title{Find SNPs with similar characteristics to the input SNPs.}
\usage{
SNPmatch_XList(database, inputSNPs, MAF, GD, DNG, LDB)
}
\arguments{
\item{database}{Data frame with information of SNPs}

\item{inputSNPs}{Character vector with the input SNPs in format chr:pos.}

\item{MAF}{Numeric value with interval limits of minor allele frequency.}

\item{GD}{Numeric value with interval limits of gene density.}

\item{DNG}{Numeric value with interval limits of distance to nearest gene.}

\item{LDB}{Numeric value with interval limits of LD buddies.}
}
\value{
Data table with the input SNPs, their matches and information about the matches.
}
\description{
Based on the information provided in the database, the function matches
the input SNPs with those SNPs in the database based on the parameters of
minor allele frequency, gene density, distance to nearest gene and
LD buddies.
}
\details{
It also writes the information on the included ("input_snps_annotated.txt")
and excluded ("input_snps_excluded.txt") SNPs depending on if they are available
on the database.

Also, at the end, if there is any SNP that does not have any match,
it will also write it in a txt ("input_snps_annotated_unmatched.txt")
and exclude it from the final output ("matched_snps_annotation.txt").

It is different to the function "SNPmatch" as this function outputs a
list with the information of the SNPs and its matches, which is also done
with the functions "listMatches" and "listMatches2". Thus, this function
is better in terms of computation time.
}
\examples{
myDatabase <- data.frame(snpID= c("10:10001753", "10:10001794"), snp_maf= c(0.074, 0.410), gene_count= c(0,0), dist_nearest_gene_snpsnap= c(98932,98891), friends_ld05= c(35,168))
myVars <- data.frame(input_snp= c("18:53093251", "12:57682956", "10:10001794"))
SNPlist <- SNPmatch_list(database = myDatabase, inputSNPs = myVars, MAF = 0.05, GD = 0.5, DNG = 0.5, LDB = 0.5)

}
