% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinSNP.R
\docType{package}
\name{twinSNP}
\alias{twinSNP}
\title{twinSNP package}
\description{
For a set of variants, it generates a vector of variants with similar characteristics of gene density, linkage disequilibrium, minor allele frequency and distance to nearest gene
}
\author{
Lucia de Hoyos \email{luciadeh@gmail.com}
}
